Add new Custom User from project Add new configs for admin Custom User Add mensage in user_not_authenticat decorator Add new Labels in Django Forms Profile Add Settings.py widget tweaks and AUTH_USER_MODEL to new custom User Model
Bugfix User Validator Info allowed duplicates info to name and matricula
Bugfix UserProfile models not set correct validators
Bugfix Login using new Custom User - default email to login
Bugfix Forms profile register info not work using incorrect atributes for fields and return error
Bugfix Mensagens error in forms atributes fields
Bugfix Teacher user to set new Custom User
Bugfix Jobs admin fieldsets not appear beneficios and tipo vaga
Bugfix If the teacher who created the job is different from the one you want to delete, you cannot allow

Change Profille html
Change Alerts to add 2 margin top

+++ djangoFULL/opportunity/apps/project/views/views.py
@@ -99,11 +99,16 @@ def search(request):
 
 @login_required
 def delete_job(request, vaga_id):
+    nome = request.user
 
     context = {}
     job = get_object_or_404(vagasEmprego, pk=vaga_id)
     context['title'] = 'deletar vaga'
     context['vaga'] = job
+
+    if job.professor.user != nome:
+        messages.error(request, "Você não pode realizar essa operação!")
+        return redirect("index")
 
     if request.method == "POST":
         job.delete()
